Small adjustements on architecture view cards and general code restructure
